+++ cpp/collections_module/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(collections_src
-    collections.cpp
-    algorithm/algorithm.cpp)
+    collections_module.cpp
+    algorithm/collections.cpp)
 
 add_query_module(collections 1 "${collections_src}")
