+++ cpp/collections_module/algorithm/collections.hpp
